backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
google-generativeai==0.3.2
PyGithub==2.1.1
slack-sdk==3.26.1
aiohttp>=3.9.0